+++ controllers/contacts/createContacts.js
@@ -1,5 +1,6 @@
 
 const handler = require('../handlers')
+const validator = require('../validators/contacts/createValidator')
 
 async function createContacts(req, res, next) {
     const contact = {
@@ -8,8 +9,8 @@ async function createContacts(req, res, next) {
     phone: req.body.phone
     };
     
-    const result = handler.schema.validate(req.body)    
-  if (result.error) {
+    const resultValidate = validator.schema.validate(req.body)    
+  if (resultValidate.error) {
     return res.status(400).send({ message: "missing required name - field"})
   } else {
   handler.addContact(contact);
